feat(controller): edit Controller to RestController

+++ src/main/java/com/space/controller/ShipController.java
@@ -2,16 +2,19 @@ package com.space.controller;
 
 import com.space.model.Ship;
 import com.space.service.ShipService;
-import com.space.service.ShipServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.util.UriComponentsBuilder;
+
 import java.util.List;
 
-//@RestController
-@Controller
-@RequestMapping("/rest")
+@RestController
+@RequestMapping("/rest/ships")
 public class ShipController {
 
     private final ShipService shipService;
@@ -21,52 +24,86 @@ public class ShipController {
         this.shipService = shipService;
     }
 
-    @GetMapping("/ships")
-    public String getShipsList(Model model) {
+    @GetMapping(value = "", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<List<Ship>> getShipsList() {
         List<Ship> ships = shipService.findAll();
-        model.addAttribute("ships", ships);
 
-        return "test";
-//        return "index";
+        if (ships == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(ships, HttpStatus.OK);
     }
 
-    @GetMapping("/ships/count")
-    public String getShipsCount() {
+    @GetMapping(value = "/count", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<Integer> getShipsCount() {
+        List<Ship> ships = shipService.findAll();
 
+        if (ships == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
 
-        return "test";
-//        return "index";
+        Integer shipCount = ships.size();
+
+        return new ResponseEntity<>(shipCount, HttpStatus.OK);
     }
 
-    @PostMapping("/ships")
-    public String createShip(Ship ship) {
+    @PostMapping(value = "", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<Ship> saveShip(@RequestBody Ship ship) {
+        HttpHeaders headers = new HttpHeaders();
+
+        if (ship == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         shipService.saveShip(ship);
 
-        return "test";
-//        return "index";
+        return new ResponseEntity<>(ship, headers, HttpStatus.OK);
     }
 
-    @GetMapping("/ships/{id}")
-    public String getShip(@PathVariable Long id) {
-        shipService.findById(id);
+    @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<Ship> getShip(@PathVariable Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
 
-        return "test";
-//        return "index";
+        Ship ship = shipService.getById(id);
+
+        if (ship == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(ship, HttpStatus.OK);
     }
 
-    @PostMapping("/ships/{id}")
-    public String updateShip(@PathVariable Long id) {
+    @PostMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<Ship> updateShip(@RequestBody @PathVariable Long id, UriComponentsBuilder builder) {
+        HttpHeaders headers = new HttpHeaders();
 
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
 
-        return "test";
-//        return "index";
+        Ship ship = shipService.getById(id);
+
+        shipService.saveShip(ship);
+
+        return new ResponseEntity<>(ship, headers, HttpStatus.OK);
     }
 
-    @DeleteMapping("/ships/{id}")
-    public String deleteShip(@PathVariable Long id) {
+    @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<Ship> deleteShip(@PathVariable Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        Ship ship = shipService.getById(id);
+
+        if (ship == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         shipService.deleteById(id);
 
-        return "test";
-//        return "index";
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
